Rework to grid*
Flexbox wasn't the best because you had to calculate the size of the
squares, which resulted in bugs, namely in firefox, where squares would
not fit inside the container. Reworked it so there is no calculation,
rather we are setting the number of columns and rows instead of size of
each individual square.